Move `FilterCriteria` to new namespace
