+++ osu.Game.Tests/NonVisual/Filtering/FilterQueryParserTest.cs
@@ -4,6 +4,7 @@
 using System;
 using NUnit.Framework;
 using osu.Game.Beatmaps;
+using osu.Game.Rulesets.Filter;
 using osu.Game.Screens.Select;
 
 namespace osu.Game.Tests.NonVisual.Filtering
