+++ osu.Game/Rulesets/Filter/FilterCriteria.cs
@@ -7,10 +7,9 @@ using System.Linq;
 using JetBrains.Annotations;
 using osu.Game.Beatmaps;
 using osu.Game.Collections;
-using osu.Game.Rulesets;
 using osu.Game.Screens.Select.Filter;
 
-namespace osu.Game.Screens.Select
+namespace osu.Game.Rulesets.Filter
 {
     public class FilterCriteria
     {
